Add a new "make travis-tests" target
This target depends on "make flake8" and calls "make test" directly so
that both flake8 and app-specific tests are performed.  This is the
testing we would like Travis to perform, (hence the name).

For this to work, flake8 is added to the requirements when
testing. This allows a direct call to "make flake8" to work as well.