Share a single sqlite3 database across each run of "make test"
This commit adds a 'TEST':{'NAME'} value within the
DATABASE['default'] setting so that "./manage.py test" will use the
same on-disk SQLite3 database as "./manage.py shell". Additionally,
this commit adds the --keepdb option to the "./manage.py test" command
so that each invocation of "make test" will reuse any database from a
previous run.

There's not any significant time savings from reusing the database
with this tiny app, but this might be a useful approach for apps. with
larger demands on the database.

Finally, if the user does need to start a new database for any reason,
this commit provides a "make distclean" target that will remove the
database. (Of course, "git clean -f -x" could also be used to do
that.)