Add models for SlowishAccount and SlowishUser
The SlowishAccount model simply has a numeric ID, (which is often
called a "tenant_id" in the Swift API). A user belongs to an account
and has a username and password as well a a generated token.

Along with the new model, we add the new migration, generated by
"./manage.py makemigrations".