Add a test for authenticating a user
The new test creates a user in the database and verifies that the
"tokens" view correctly authenticates (with the correct password) and
also that it fails to authenticate (when the password is incorrect).

This new test is expected to fail currently, (since we do not yet have
code to verify a username and password in the tokens view yet).