Implement verification of account.id, username, and password in tokens view.
This fixes the test that was recently added. In the case of an
authorized user this view uses a template to return a json-formatted
response that gives the URL of the endpoint needed for future requests
and the token that will be needed for those requests.

(Note: The endpoint URL returned is correct, but following it up will
not currently work since this app does not yet have an implementation
of the views necessary for the subsequent requests.)