Augment testing to improve code coverage
The recently-added container() view had some branches in it that
were not being covered by the test suite, (things like issuing a PUT
to create a container just after having issued an identical PUT to
create it in the first place).

All of the tests added here are expected to pass already. And with
this commit the reported test coverage is back up to 100%.