Implement correct list of available containers
It's simple enough to loop over the containers for the current account
and package them up in a JSON list.

The JSON API is a bit paranoid about the top-level object in JSON
being an array, (apparently there have been browser bugs in the past),
so it requires a "safe=False" to do this. This commit includes a
comment to justify relaxing Django's paranoia here.