Add test for 'marker' and 'endmarker' in query string
The swift client uses these to determine when it has reached the end
of the list. So this implementation cannot survive even a "swift list"
invocation unless it supports these parameters.

This new test is expected to fail currently, since the code to support
these parameters is not yet implemented.