Add a SlowishFile model with a migration, a view, and a test
This allows a file to be added to a container. Currently, the
implementation of a file is extremely minimal---beyond the key to the
containing container, it contains only a path. It does not yet contain
any type, (such as "application/directory", or "text/plain"), nor does
it contain any content.