Test that we can authenticate with a "tenantName" entry
Commit a8f02301ff6fd5e938d5708bd60b43add016bdf8 added support for
authenticating with an entry in the "auth" dictionary of "tenantName"
instead of "tenantId". This had been untested until now.

With this commit, "make coverage" now reports that 100% of the lines
of code in this app are executed at least once by "make
test". Obviously, that doesn't guarantee that the app is correct in
all cases, but it's a nice bar to have met anyway.