Use more efficient and idiomatic code to implement markers
Previously, the code here was using exceptions for flow control. It's
cleaner, (and less total code), to simply say, "if value in
QueryDict:".

Additionally, the code had been using a lambda to walk the results of
the Django QuerySet to filter down to the elements desired. In this
commit, we instead chain an additional filter() call onto the QuerySet
itself. This means that the filtering will happen entirely within the
database and no results outside the range desired will ever even be
returned by the time the QuerySet is evaluated.